# generatedOpenapi.yaml
openapi: 3.1.0
info:
  title: Bank API
  description: "REST API for a Eagle Bank which allows a user to create, fetch, update\
    \ and delete a bank account and deposit or withdraw money from the account."
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local Development Server
...
paths:
  /v1/users:
    post:
      tags:
      - user
      summary: Create a new user
      description: Create a new user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequestDto"
        required: true
      responses:
        "201":
          description: User has been created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: Invalid details supplied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponseDto"
        "409":
          description: User with email already exists
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: An unexpected error occurred
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
  /v1/users/{userId}:
    get:
      tags:
      - user
      summary: Fetch user by ID.
      description: Fetch user by ID.
      operationId: fetchUserByID
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: The user details
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: The request didn't supply all the necessary data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponseDto"
        "401":
          description: Access token is missing or invalid
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: The user is not allowed to access the transaction
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: User was not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: An unexpected error occurred
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - bearerAuth: []
    delete:
      tags:
      - user
      summary: Delete user by ID.
      description: Delete user by ID.
      operationId: deleteUserByID
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: The user has been deleted
        "400":
          description: The request didn't supply all the necessary data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponseDto"
        "401":
          description: Access token is missing or invalid
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: The user is not allowed to access the transaction
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: User was not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: A user cannot be deleted when they are associated with a bank
            account
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: An unexpected error occurred
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - bearerAuth: []
    patch:
      tags:
      - user
      summary: Update user by ID.
      description: Update user by ID.
      operationId: updateUserByID
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
        required: true
      responses:
        "200":
          description: The updated user details
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: The request didn't supply all the necessary data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponseDto"
        "401":
          description: Access token is missing or invalid
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: The user is not allowed to access the transaction
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: User was not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: An unexpected error occurred
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - bearerAuth: []
...
components:
  schemas:
    Address:
      type: object
      properties:
        line1:
          type: string
          maxLength: 100
          minLength: 0
        line2:
          type: string
          maxLength: 100
          minLength: 0
        line3:
          type: string
          maxLength: 100
          minLength: 0
        town:
          type: string
          maxLength: 50
          minLength: 0
        county:
          type: string
          maxLength: 50
          minLength: 0
        postcode:
          type: string
          maxLength: 20
          minLength: 0
      required:
      - county
      - line1
      - postcode
      - town
    CreateUserRequestDto:
      type: object
      properties:
        email:
          type: string
          maxLength: 100
          minLength: 0
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
        name:
          type: string
          maxLength: 100
          minLength: 0
        address:
          $ref: "#/components/schemas/Address"
        phoneNumber:
          type: string
          minLength: 1
          pattern: "^\\+[1-9]\\d{1,14}$"
      required:
      - address
      - email
      - name
      - password
      - phoneNumber
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          $ref: "#/components/schemas/Address"
        phoneNumber:
          type: string
        email:
          type: string
        createdTimestamp:
          type: string
          format: date-time
        updatedTimestamp:
          type: string
          format: date-time
    BadRequestErrorResponseDto:
      type: object
      properties:
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetailDto"
    ErrorDetailDto:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        type:
          type: string
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
...
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          maxLength: 100
          minLength: 0
        name:
          type: string
          maxLength: 100
          minLength: 0
        address:
          $ref: "#/components/schemas/Address"
        phoneNumber:
          type: string
          pattern: "^\\+[1-9]\\d{1,14}$"
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Authentication token (Bearer Token)
      scheme: bearer
      bearerFormat: JWT